CI: Drop use of uv in functional test.
While UV is great, its reliance on specific Python interpreter
versions does not work well when testing at the bleeding edge of
distro development.

As a concrete example, at this point in time it is not possible to
use uv on Ubuntu Plucky Puffin as it will fail to find development
headers for Python 3.12 as only Python 3.13 is available in the
system.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>